requirements.txt
pyfrc
wpilib
robotpy-ctre
robotpy-navx

marsutils ~= 0.2